perf(core,render,test): Tier 5 — branch hints + EVIL docs + vkWaitForFences 10ms + InputAction mask UB fix + shadow splits tests + shadow benchmark
Mixed grab-bag of perf polish + correctness fixes
from the Tier 5 backlog in `decisions.md §777`.
Six sub-tasks, all behind the same ctest gate.

**1. `[[likely]] / [[unlikely]]` annotations in `IsSceneChunkVisible`.**
Branch hints on the early-out paths (empty-AABB
cull, near-plane cull, max-distance cull, the 4
side-plane culls). The hot cull function is called
per chunk per frame; the branch predictor benefits
even on debug builds where the hint is a
`__builtin_expect` wrapper. `src/render/SceneResources.hpp`.

**2. `// EVIL:` comments on magic numbers.**
Per `legacy/docs/philosophy/01_foundation/04_evil-hacks-philosophy.md`,
every hack must be marked. `BuildSunShadowCascadeSplits`
gets comments on the `0.0001f` minimum split step
and the `1.0f` cascade depth floor. The frustum
cull benchmark fixture gets comments on the
`tan(39.7°)` / `tan(50°)` FOV tangents and the
`0.5f` near plane.

**3. `vkWaitForFences` tight latency (2 sites).**
Per `legacy/docs/philosophy/01_foundation/01_manifesto.md`
'p99 latency, не средний' — the engine polls for
the previous frame's GPU work with a 10 ms budget
and falls back to blocking only if the GPU is
behind. The fallback is wrapped in
`PV_PROFILE_ZONE_N("...GPUStallFallback")` for
Tracy. The 10 ms constant is `// EVIL:`-marked as
a hard-coded budget (a future refactor should drive
it from `Configuration::frameFenceTimeoutNs`).
`src/app/FramePreparation.cpp:115`,
`src/render/Renderer.cpp:231`.

**4. `InputAction` bit-mask overflow fix (real bug).**
The 58-action `InputAction` inventory exceeded the
32-bit replay mask capacity at some point in the
music / track control additions. The pre-Tier-5
`uint32_t actionDownMask / actionPressedMask`
silently lost bits 32..57 and invoked **undefined
behavior** on the `1u << actionIndex` shift (UB
on `shift >= 32` for `uint32_t`). Nobody noticed
because the most-used actions are all in the low
bits and the high-bit actions are debug-only.

Fix:
- Widen `InputReplayFrame::actionDownMask /
  actionPressedMask` to `uint64_t`.
- Widen `GetInputActionDownMask / Pressed`
  return type and `ApplyInputActionSnapshot`
  parameter type to `uint64_t`.
- Replace `1u << actionIndex` with
  `1ull << actionIndex` in `InputActions.cpp`
  + the 3 test sites in `VoxelWorldTests.cpp`.
- Bump on-disk replay format version
  `kInputReplayVersion` 2 → 3.
- Add `static_assert(kInputActionCount <= 64)`
  in `core/Types.hpp` to make adding a 65th
  action a compile-time failure.

`v1` and `v2` replay files remain loadable (the
per-frame integer stream is binary-compatible;
`stream >> uint64_t` zero-extends a v2 `uint32_t`
token correctly).

**5. Tests for `BuildSunShadowCascadeSplits`.**
New ctest target `ProjectVSunShadowCascadeSplitsTests`
with 7 sub-tests:
- happy path (0.1→128, default lambda)
- `minDepthStep` regression check (0.01→1.0)
- `splitLambda=0` → pure uniform
- `splitLambda=1` → pure logarithmic
- defensive defaults for NaN / non-finite inputs
- `splitLambda` clamp `< 0` and `> 1`
- degenerate `farPlane <= nearPlane`

CPU-only, no Vulkan, no asset pipeline. `tests/SunShadowCascadeSplitsTests.cpp`.

**6. `ShadowProjectionBenchmark`.**
Google Benchmark harness for the per-frame shadow
projection hot path. Two benchmarks:
- `BM_BuildSunShadowProjection` — 1456 ns/op
  (per-frame shadow VP, 24×16×24 voxel world with
  Y=0 floor)
- `BM_BuildSunShadowCascadeSplits` — 116 ns/op
  (per-frame cascade splits)

Output is a numeric `ns/op + stddev` for perf
regression tracking per `decisions.md §777`. The
total shadow projection cost is ~1.5 µs/frame
(~0.01% of a 16.6 ms 60 Hz budget) — not a
bottleneck today, but the baseline number is
captured for future refactors.

**Test results.** 12/12 ctest pass (was 11/11;
added `ProjectVSunShadowCascadeSplitsTests`).
`bin/ProjectV` runs cleanly with all changes.

Refs: AGENTS.md §7.2.5, AGENTS.md §7.2.7,
agent/memory.md §10.11, agent/memory.md §11.3,
decisions.md §777

+++ src/app/InputActions.cpp
@@ -258,31 +258,34 @@ bool ConsumeInputActionPressed(
 	return wasPressed;
 }
 
-uint32_t GetInputActionDownMask(const InputState &input)
+// **Tier 5 (`2026-06-13`).** `uint64_t` return type
+// and `1ull << actionIndex` shift (the `1u` 32-bit
+// shift with `actionIndex >= 32` was UB).
+uint64_t GetInputActionDownMask(const InputState &input)
 {
-	uint32_t mask = 0;
+	uint64_t mask = 0;
 	for (size_t actionIndex = 0; actionIndex < input.actions.size(); ++actionIndex) {
 		const InputAction action = static_cast<InputAction>(actionIndex);
 		if (!IsInputActionReplayRecordable(action) || !input.actions[actionIndex].down) {
 			continue;
 		}
 
-		mask |= 1u << actionIndex;
+		mask |= 1ull << actionIndex;
 	}
 
 	return mask;
 }
 
-uint32_t GetInputActionPressedMask(const InputState &input)
+uint64_t GetInputActionPressedMask(const InputState &input)
 {
-	uint32_t mask = 0;
+	uint64_t mask = 0;
 	for (size_t actionIndex = 0; actionIndex < input.actions.size(); ++actionIndex) {
 		const InputAction action = static_cast<InputAction>(actionIndex);
 		if (!IsInputActionReplayRecordable(action) || !input.actions[actionIndex].pressed) {
 			continue;
 		}
 
-		mask |= 1u << actionIndex;
+		mask |= 1ull << actionIndex;
 	}
 
 	return mask;
@@ -290,15 +293,15 @@ uint32_t GetInputActionPressedMask(const InputState &input)
 
 void ApplyInputActionSnapshot(
 	InputState &input,
-	const uint32_t downMask,
-	const uint32_t pressedMask)
+	const uint64_t downMask,
+	const uint64_t pressedMask)
 {
 	input.lastMoveUpPressedTimestampNs = 0;
 	for (size_t actionIndex = 0; actionIndex < input.actions.size(); ++actionIndex) {
 		const InputAction action = static_cast<InputAction>(actionIndex);
 		const bool isRecordable = IsInputActionReplayRecordable(action);
-		const bool isDown = isRecordable && (downMask & 1u << actionIndex) != 0u;
-		const bool isPressed = isRecordable && (pressedMask & 1u << actionIndex) != 0u;
+		const bool isDown = isRecordable && (downMask & 1ull << actionIndex) != 0u;
+		const bool isPressed = isRecordable && (pressedMask & 1ull << actionIndex) != 0u;
 		input.actions[actionIndex].down = isDown;
 		input.actions[actionIndex].pressed = isPressed;
 		input.bindings[actionIndex].downStates = {};
